+++ test/typeclass/Coproduct.ts
@@ -1,4 +1,3 @@
-import { pipe } from "@fp-ts/core/Function"
 import * as O from "@fp-ts/core/Option"
 import * as _ from "@fp-ts/core/typeclass/Coproduct"
 import * as U from "../util"
@@ -6,13 +5,13 @@ import * as U from "../util"
 describe("Coproduct", () => {
   it("getMonoid", () => {
     const M = _.getMonoid(O.Alternative)<unknown, never, never, number>()
-    U.deepStrictEqual(pipe(O.none(), M.combine(O.none())), O.none())
-    U.deepStrictEqual(pipe(O.some(1), M.combine(O.none())), O.some(1))
-    U.deepStrictEqual(pipe(O.none(), M.combine(O.some(2))), O.some(2))
-    U.deepStrictEqual(pipe(O.some(1), M.combine(O.some(2))), O.some(1))
+    U.deepStrictEqual(M.combine(O.none(), O.none()), O.none())
+    U.deepStrictEqual(M.combine(O.some(1), O.none()), O.some(1))
+    U.deepStrictEqual(M.combine(O.none(), O.some(2)), O.some(2))
+    U.deepStrictEqual(M.combine(O.some(1), O.some(2)), O.some(1))
 
-    U.deepStrictEqual(pipe(M.empty, M.combine(O.none())), O.none())
-    U.deepStrictEqual(pipe(M.empty, M.combine(O.some(2))), O.some(2))
-    U.deepStrictEqual(pipe(O.some(1), M.combine(M.empty)), O.some(1))
+    U.deepStrictEqual(M.combine(M.empty, O.none()), O.none())
+    U.deepStrictEqual(M.combine(M.empty, O.some(2)), O.some(2))
+    U.deepStrictEqual(M.combine(O.some(1), M.empty), O.some(1))
   })
 })